chore: rename preparation -> shape_inference
