+++ wonnx-preprocessing/src/lib.rs
@@ -3,7 +3,7 @@ use wonnx::utils::InputTensor;
 
 pub mod constant_folding;
 pub mod image;
-pub mod preparation;
+pub mod shape_inference;
 pub mod text;
 
 pub enum Tensor {
